feat: go router added, ui changes, code refactor
