+++ lib/core/utils/colors.dart
@@ -1,0 +1,8 @@
+class CatBreedsColors {
+  //prevents instantiation and extension
+  CatBreedsColors._();
+
+  static const int primary = 0xFF6068A6;
+  static const int secondary = 0xFFF7EFFB;
+  static const int backgroundColor = 0xFFD9D5EA;
+}
