+++ lib/main_production.dart
@@ -1,5 +1,5 @@
-import 'package:catbreeds/modules/presentation/app/app.dart';
 import 'package:catbreeds/bootstrap.dart';
+import 'package:catbreeds/modules/presentation/app/app.dart';
 
 void main() {
   bootstrap(() => const App());
