+++ pubspec.yaml
@@ -48,3 +48,4 @@ flutter:
   generate: true
   assets:
     - assets/icon/
+    - assets/images/
